+++ apps/web/src/components/schedule/DateDropdown/DateDropdown.css.ts
@@ -1,0 +1,6 @@
+import { style } from '@vanilla-extract/css';
+import { vars } from '@repo/theme';
+
+export const contentStyle = style({
+  padding: `${vars.space[20]}`,
+});
